requirements.txt
opencv-python==4.5.3.56
pyzbar==0.1.8
qrcode
pillow==8.3.2